requirements.txt
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.13.4
python-dotenv~=1.1.0
sentence-transformers~=4.1.0
torch~=2.7.1
transformers~=4.52.4
pillow~=11.2.1
openai~=1.84.0
torchvision~=0.22.1
pymilvus~=2.5.10
streamlit~=1.45.1
datasets~=3.6.0
ragas~=0.2.15
numpy~=2.2.6
scikit-learn~=1.6.1